Make index operations tolerant of Windows paths. (#2)
Make index operations tolerant of Windows paths.

In Windows, the '\' is used for path separation. In the Go standard
library, the `path/filepath` package seems to handle this correctly, but
the `path` package seems to get it wrong. In all of the relevant places
this has been swapped out, and a few tests have been added.